backend/requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
pydantic
python-multipart
python-jose[cryptography]
passlib[bcrypt]
easyocr
opencv-python-headless
numpy
